move to mutable state, dirty-checking and renderloop
This commit makes the code quite simpler removing all the callback
passing and uses a mutable opaque type as the state. Edit operations on
the state mark it as dirty, causing a redraw during a rAF render loop.

On my device this results in noticeably better response times by the canvas
when drawing new lines.