add color schemes support
Both AnkiDroid and Anki have a Night Mode which adds an element with
a `night_mode` class to the DOM.

With this patch, if the new colorScheme option is set to auto (which is
the default value) we query for the aforementioned element to decide
which color scheme to render.

The user can also customize single colors in the scheme, and add their
own scheme through AnkiCanvasOptions.